Cambio de nombres de HistoriaActividad a HistoryGame
