+++ src/components/HistoryGame/HistoryGame.css
@@ -1,0 +1,34 @@
+.historia-container {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 30px;
+  }
+  
+  .historia-content {
+    max-width: 600px;
+    border-radius: 12px;
+    text-align: center;
+  }
+  
+  .historia-title {
+    color: #9d4528;
+    font-size: 22px;
+    margin-bottom: 20px;
+    font-weight: bold;
+  }
+  
+  .historia-imagen {
+    width: 100%;
+    max-width: 300px;
+    margin: 20px auto;
+  }
+  
+  .historia-texto {
+    font-size: 18px;
+    line-height: 1.5;
+    color: #333;
+    margin-bottom: 30px;
+  }
+
